Implemented recommended changes
