+++ src/gameplay/game.ts
@@ -13,7 +13,7 @@ import { isTO } from '../modsadmins/tournamentOrganizers';
 import { isDev } from '../modsadmins/developers';
 import { modOrTOString } from '../modsadmins/modOrTO';
 
-import { roomCreationTypeEnum, getRoomTypeFromString } from './enums/roomCreationType';
+import { roomCreationTypeEnum, getRoomTypeFromString } from './roomTypes';
 import { gameModeObj } from './gameModes';
 
 class Game extends Room {
@@ -23,8 +23,6 @@ class Game extends Room {
   phase = 'pickingTeam';
   missionHistory = [];
   roomCreationType : roomCreationTypeEnum;
-
-  // Add const gameType = "Ranked"/"Unranked"/"Custom";
 
   // TODO This shouldn't be here! Should be in Assassin file.
   startAssassinationTime: Date;
@@ -135,8 +133,6 @@ class Game extends Room {
     this.hammer = 0;
     this.missionNum = 0;
     this.pickNum = 0;
-
-    // this.gameType = gameType
 
     this.numFailsHistory = [];
     this.proposedTeam = [];
@@ -1345,15 +1341,17 @@ class Game extends Room {
     //FR2 - Rating system 1 - record games for rating updates   
     // This data will be used to calculate ratings at end of rating period.
     // The intended workflow is games -> ratingPeriodGameRecord -> (at end of rating preiod) Ratings Calculator
-
+    assert.equal(this.spyUsernames.length + this.resistanceUsernames.length, this.playersInGame.length, 
+      'Number of players should be number of resistance players + number of spy players'
+    );
     if (this.roomCreationType === roomCreationTypeEnum.RANKED_QUEUE) {
       const ratingRecordToStore = {
-        timeGameStarted: this.startGameTime,
-        timeGameFinished: timeFinished,
+        // timeGameStarted: this.startGameTime, // Removed 
+        timeGameFinished: timeFinished, // For logging only
         winningTeam: this.winner,
         spyTeam: this.spyUsernames,
         resistanceTeam: this.resistanceUsernames,
-        numberOfPlayers: this.playersInGame.length,
+        // numberOfPlayers: this.playersInGame.length, // Assert above
         roomCreationType: this.roomCreationType,
       };
 
@@ -1388,9 +1386,9 @@ class Game extends Room {
         provisionalGame = true;
       }
 
-// // TODO Disable the ratings since rating will now be calculated at the end of the rating period
+// TODO Disable the ratings since rating will now be calculated at the end of the rating period
 
-      /* // calculate team 1v1 elo adjustment
+      // calculate team 1v1 elo adjustment
       const teamResChange = this.calculateResistanceRatingChange(
         this.winner,
         provisionalGame,
@@ -1401,10 +1399,10 @@ class Game extends Room {
       const indResChange =
         Math.round((teamResChange / this.resistanceUsernames.length) * 10) / 10;
       const indSpyChange =
-        Math.round((teamSpyChange / this.spyUsernames.length) * 10) / 10; */
+        Math.round((teamSpyChange / this.spyUsernames.length) * 10) / 10;
 
       // if we're in a ranked game show the elo adjustments
-      /* if (this.ranked) {
+      if (this.ranked) {
         // Get the old player ratings (with usernames) for use in provisional calculations.
         const oldPlayersInfo = this.playersInGame.map((soc) => {
           const data = {};
@@ -1485,7 +1483,7 @@ class Game extends Room {
             }
           }
         });
-      } */
+      }
 
       this.playersInGame.forEach((player) => {
         User.findById(player.userId)
@@ -1518,8 +1516,8 @@ class Game extends Room {
               if (this.ranked) {
                 foundUser.totalRankedGamesPlayed += 1;
 
-                // Disabling since it is not being calculated here
-                /* if (foundUser.ratingBracket === 'unranked') {
+                // TODO Disable it when elo is no longer being calculated here
+                if (foundUser.ratingBracket === 'unranked') {
                   foundUser.playerRating = player.request.user.playerRating;
                 } else {
                   if (player.alliance === 'Resistance') {
@@ -1527,7 +1525,7 @@ class Game extends Room {
                   } else if (player.alliance === 'Spy') {
                     foundUser.playerRating += indSpyChange;
                   }
-                } */
+                } 
               }
 
               // checks that the var exists
